\name{smfsb-package}
\alias{smfsb}
\alias{smfsb2e}
\alias{SMfSB}
\alias{SMfSB2e}

\docType{package}

\title{Stochastic modelling for systems biology, 2nd edition}

\description{
This package contains code and data for modelling and simulation of stochastic kinetic biochemical network models. It contains the code and data associated with the second edition of the book Stochastic Modelling for Systems Biology, to be published by Chapman & Hall/CRC Press late 2011.
}

\author{
Maintainer: Darren Wilkinson <d.j.wilkinson@ncl.ac.uk>
}
\references{See \url{http://www.staff.ncl.ac.uk/d.j.wilkinson/smfsb/2e/}}

\keyword{package}

}


