\name{spnModels}
\alias{spnModels}
\alias{LV}
\alias{ID}
\alias{BD}
\alias{Dimer}
\docType{data}
\title{Example SPN models}
\description{
Collection of example SPN models. Includes \code{LV}, a Lotka--Volterra
model, \code{ID}, an immigration--death process, \code{BD}, a birth--death
process and \code{Dimer}, a simple dimerisation kinetics model.
}
\usage{data(spnModels)}
\format{Each model is a list, with components \code{Pre}, \code{Post},
  and \code{h}. See \code{\link{StepGillespie}} for further details, and
  examples of use.}
\keyword{smfsb}
\keyword{data}
\keyword{models}
\keyword{datasets}

