\name{smfsb-package}
\alias{smfsb}
\alias{smfsb2e}
\alias{smfsb3e}
\alias{SMfSB}
\alias{SMfSB2e}
\alias{SMfSB3e}

\docType{package}

\title{Stochastic Modelling for Systems Biology}

\description{
This package contains code and data for modelling and simulation of
stochastic kinetic biochemical network models. It contains the code and
data associated with the second and third editions of the book Stochastic Modelling
for Systems Biology, published by Chapman & Hall/CRC Press.
}

\author{
Maintainer: Darren Wilkinson <darren.wilkinson@ncl.ac.uk>
}
\references{See \url{http://www.staff.ncl.ac.uk/d.j.wilkinson/smfsb/}
  or \url{https://github.com/darrenjw/smfsb}
for further details.}

\keyword{package}



