\name{smfsb-package}
\alias{smfsb}
\alias{smfsb2e}
\alias{SMfSB}
\alias{SMfSB2e}

\docType{package}

\title{Stochastic Modelling for Systems Biology, second edition}

\description{
This package contains code and data for modelling and simulation of
stochastic kinetic biochemical network models. It contains the code and
data associated with the second edition of the book Stochastic Modelling
for Systems Biology, published by Chapman & Hall/CRC Press, November 2011.
}

\author{
Maintainer: Darren Wilkinson <darren.wilkinson@ncl.ac.uk>
}
\references{See \url{http://www.staff.ncl.ac.uk/d.j.wilkinson/smfsb/2e/}
  or \url{http://tinyurl.com/smfsb2e}
for further details.}

\keyword{package}



